api/requirements.txt
bcrypt==3.1.7
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
fastapi==0.44.1
google-api-python-client==1.7.11
google-auth==1.10.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
h11==0.8.1
httplib2==0.15.0
httptools==0.0.13
idna==2.8
oauthlib==3.1.0
passlib==1.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pydantic==1.2
PyJWT==1.7.1
python-multipart==0.0.5
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
starlette==0.12.9
uritemplate==3.0.1
urllib3==1.25.7
uvicorn==0.10.8
uvloop==0.14.0
websockets==8.1
